+++ moncic/ci.py
@@ -68,14 +68,14 @@ class MoncicCommand(Command):
         parser = super().make_subparser(subparsers)
         parser.add_argument("-I", "--imagedir", action="store",
                             help="path to the directory that contains container images."
-                                 " Default: from configuration file, or ./images")
+                                 " Default: from configuration file, or /var/lib/machines")
         return parser
 
     def __init__(self, args):
         super().__init__(args)
         self.moncic = Moncic()
         if self.args.imagedir:
-            self.moncic.config.imagedir = self.args.imagedir
+            self.moncic.set_imagedir(self.args.imagedir)
 
 
 class CI(MoncicCommand):
@@ -100,17 +100,16 @@ class CI(MoncicCommand):
         return parser
 
     def run(self):
-        with self.moncic:
-            system = self.moncic.create_system(self.args.system)
+        with self.moncic.images() as images:
+            system = images.create_system(self.args.system)
             with checkout(system, self.args.repo, branch=self.args.branch) as srcdir:
                 container = system.create_container(config=ContainerConfig(ephemeral=True, workdir=srcdir))
                 if self.args.build_style:
                     builder = Builder.create(self.args.build_style, container)
                 else:
                     builder = Builder.detect(container)
-                with self.moncic.privs.root():
-                    with container:
-                        res = container.run_callable(builder.build)
+                with container:
+                    res = container.run_callable(builder.build)
                 return res.returncode
 
 
@@ -156,8 +155,8 @@ class ImageActionCommand(MoncicCommand):
 
     @contextlib.contextmanager
     def container(self):
-        with self.moncic:
-            system = self.moncic.create_system(self.args.system)
+        with self.moncic.images() as images:
+            system = images.create_system(self.args.system)
             if not system.is_bootstrapped():
                 raise Fail(f"{system.name!r} has not been bootstrapped")
 
@@ -177,9 +176,8 @@ class ImageActionCommand(MoncicCommand):
                 if self.args.bind_ro:
                     config.bind_ro = self.args.bind_ro
 
-                with self.moncic.privs.root():
-                    with system.create_container(config=config) as container:
-                        yield container
+                with system.create_container(config=config) as container:
+                    yield container
 
 
 class Shell(ImageActionCommand):
@@ -250,34 +248,33 @@ class Bootstrap(MoncicCommand):
         return parser
 
     def run(self):
-        with self.moncic:
+        with self.moncic.images() as images:
             if not self.args.systems:
-                systems = self.moncic.list_images()
+                systems = images.list_images()
             else:
                 systems = self.args.systems
 
-            systems = self.moncic.add_dependencies(systems)
+            systems = images.add_dependencies(systems)
 
             for name in systems:
-                system = self.moncic.create_system(name)
-                with self.moncic.privs.root():
-                    if self.args.recreate and os.path.exists(system.path):
-                        system.remove()
+                system = images.create_system(name)
+                if self.args.recreate and os.path.exists(system.path):
+                    system.remove()
 
-                    if not os.path.exists(system.path):
-                        log.info("%s: bootstrapping subvolume", name)
-                        try:
-                            system.bootstrap()
-                        except Exception:
-                            log.critical("%s: cannot create image", name, exc_info=True)
-                            return 5
-
-                    log.info("%s: updating subvolume", name)
+                if not os.path.exists(system.path):
+                    log.info("%s: bootstrapping subvolume", name)
                     try:
-                        system.update()
+                        system.bootstrap()
                     except Exception:
-                        log.critical("%s: cannot update image", name, exc_info=True)
-                        return 6
+                        log.critical("%s: cannot create image", name, exc_info=True)
+                        return 5
+
+                log.info("%s: updating subvolume", name)
+                try:
+                    system.update()
+                except Exception:
+                    log.critical("%s: cannot update image", name, exc_info=True)
+                    return 6
 
 
 class Update(MoncicCommand):
@@ -292,9 +289,9 @@ class Update(MoncicCommand):
         return parser
 
     def run(self):
-        with self.moncic:
+        with self.moncic.images() as images:
             if not self.args.systems:
-                systems = self.moncic.list_images()
+                systems = images.list_images()
             else:
                 systems = self.args.systems
 
@@ -302,24 +299,20 @@ class Update(MoncicCommand):
             count_failed = 0
 
             for name in systems:
-                system = self.moncic.create_system(name)
+                system = images.create_system(name)
 
                 if not os.path.exists(system.path):
                     continue
 
                 log.info("%s: updating subvolume", name)
-                with self.moncic.privs.root():
-                    try:
-                        system.update()
-                        count_ok += 1
-                    except Exception:
-                        log.critical("%s: cannot update image", name, exc_info=True)
-                        count_failed += 1
+                try:
+                    system.update()
+                    count_ok += 1
+                except Exception:
+                    log.critical("%s: cannot update image", name, exc_info=True)
+                    count_failed += 1
 
             log.info("%d images successfully updated", count_ok)
-
-            self.moncic.deduplicate()
-            self.moncic.maybe_trim_image_file()
 
             if count_failed:
                 log.error("%d images failed to update", count_failed)
@@ -340,13 +333,12 @@ class Remove(MoncicCommand):
         return parser
 
     def run(self):
-        with self.moncic:
+        with self.moncic.images() as images:
             for name in self.args.systems:
-                system = self.moncic.create_system(name)
+                system = images.create_system(name)
                 if not os.path.exists(system.path):
                     continue
-                with self.moncic.privs.root():
-                    system.remove()
+                system.remove()
 
 
 class RowOutput:
@@ -409,9 +401,9 @@ class Images(MoncicCommand):
                     TextColumn("Boostrapped"),
                     TextColumn("Path"))
 
-        with self.moncic:
-            for name in self.moncic.list_images():
-                system = self.moncic.create_system(name)
+        with self.moncic.images() as images:
+            for name in images.list_images():
+                system = images.create_system(name)
                 output.add_row((name, system.distro.name, "yes" if system.is_bootstrapped() else "no", system.path))
         output.flush()
 
@@ -447,5 +439,5 @@ class Dedup(MoncicCommand):
     Deduplicate disk usage in image directories
     """
     def run(self):
-        with self.moncic:
-            self.moncic.deduplicate()
+        with self.moncic.images() as images:
+            images.deduplicate()